refactor: move AccessControl to Endpoint
