+++ packages/zwave-js/src/lib/node/endpoint-mixins/index.ts
@@ -1,0 +1,3 @@
+import { AccessControlMixin } from "./05_AccessControl.js";
+
+export class EndpointMixins extends AccessControlMixin {}
